+++ prawcore/__init__.py
@@ -2,6 +2,18 @@
 
 import logging
 
+from ._async.auth import (
+    AsyncAuthorizer,
+    AsyncDeviceIDAuthorizer,
+    AsyncImplicitAuthorizer,
+    AsyncReadOnlyAuthorizer,
+    AsyncScriptAuthorizer,
+    AsyncTrustedAuthenticator,
+    AsyncUntrustedAuthenticator,
+)
+from ._async.requestor import AsyncRequestor
+from ._async.sessions import AsyncSession
+from ._async.sessions import session as async_session
 from .auth import (
     Authorizer,
     DeviceIDAuthorizer,