+++ prawcore/_async/auth.py
@@ -4,21 +4,21 @@ from __future__ import annotations
 
 import time
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Any, Callable
+from typing import TYPE_CHECKING, Any, Awaitable, Callable
 
 from niquests import Request
 from niquests.status_codes import codes
 
-from . import const
-from .exceptions import InvalidInvocation, OAuthException, ResponseException
+from .. import const
+from ..exceptions import InvalidInvocation, OAuthException, ResponseException
 
 if TYPE_CHECKING:
     from niquests.models import Response
 
-    from prawcore.requestor import Requestor
+    from .requestor import AsyncRequestor
 
 
-class BaseAuthenticator(ABC):
+class AsyncBaseAuthenticator(ABC):
     """Provide the base authenticator object that stores OAuth2 credentials."""
 
     @abstractmethod
@@ -27,18 +27,18 @@ class BaseAuthenticator(ABC):
 
     def __init__(
         self,
-        requestor: Requestor,
+        requestor: AsyncRequestor,
         client_id: str,
         redirect_uri: str | None = None,
     ):
         """Represent a single authentication to Reddit's API.
 
-        :param requestor: An instance of :class:`.Requestor`.
+        :param requestor: An instance of :class:`.AsyncRequestor`.
         :param client_id: The OAuth2 client ID to use with the session.
         :param redirect_uri: The redirect URI exactly as specified in your OAuth
             application settings on Reddit. This parameter is required if you want to
             use the :meth:`~.Authorizer.authorize_url` method, or the
-            :meth:`~.Authorizer.authorize` method of the :class:`.Authorizer` class
+            :meth:`~.Authorizer.authorize` method of the :class:`.AsyncAuthorizer` class
             (default: ``None``).
 
         """
@@ -46,10 +46,10 @@ class BaseAuthenticator(ABC):
         self.client_id = client_id
         self.redirect_uri = redirect_uri
 
-    def _post(
+    async def _post(
         self, url: str, success_status: int = codes["ok"], **data: Any
     ) -> Response:
-        response = self._requestor.request(
+        response = await self._requestor.request(
             "post",
             url,
             auth=self._auth(),
@@ -82,17 +82,15 @@ class BaseAuthenticator(ABC):
 
         :raises: :class:`.InvalidInvocation` if ``redirect_uri`` is not provided, if
             ``implicit`` is ``True`` and an authenticator other than
-            :class:`.UntrustedAuthenticator` is used, or ``implicit`` is ``True`` and
-            ``duration`` is ``"permanent"``.
+            :class:`.AsyncUntrustedAuthenticator` is used, or ``implicit`` is ``True``
+            and ``duration`` is ``"permanent"``.
 
         """
         if self.redirect_uri is None:
             msg = "redirect URI not provided"
             raise InvalidInvocation(msg)
-        if implicit and not isinstance(self, UntrustedAuthenticator):
-            msg = (
-                "Only UntrustedAuthenticator instances can use the implicit grant flow."
-            )
+        if implicit and not isinstance(self, AsyncUntrustedAuthenticator):
+            msg = "Only AsyncUntrustedAuthenticator instances can use the implicit grant flow."
             raise InvalidInvocation(msg)
         if implicit and duration != "temporary":
             msg = "The implicit grant flow only supports temporary access tokens."
@@ -110,7 +108,7 @@ class BaseAuthenticator(ABC):
         request = Request("GET", url, params=params)
         return request.prepare().url
 
-    def revoke_token(self, token: str, token_type: str | None = None):
+    async def revoke_token(self, token: str, token_type: str | None = None):
         """Ask Reddit to revoke the provided token.
 
         :param token: The access or refresh token to revoke.
@@ -123,18 +121,18 @@ class BaseAuthenticator(ABC):
         if token_type is not None:
             data["token_type_hint"] = token_type
         url = self._requestor.reddit_url + const.REVOKE_TOKEN_PATH
-        self._post(url, **data)
+        await self._post(url, **data)
 
 
-class BaseAuthorizer:
+class AsyncBaseAuthorizer:
     """Superclass for OAuth2 authorization tokens and scopes."""
 
-    AUTHENTICATOR_CLASS: tuple | type = BaseAuthenticator
+    AUTHENTICATOR_CLASS: tuple | type = AsyncBaseAuthenticator
 
-    def __init__(self, authenticator: BaseAuthenticator):
+    def __init__(self, authenticator: AsyncBaseAuthenticator):
         """Represent a single authorization to Reddit's API.
 
-        :param authenticator: An instance of :class:`.BaseAuthenticator`.
+        :param authenticator: An instance of :class:`.AsyncBaseAuthenticator`.
 
         """
         self._authenticator = authenticator
@@ -146,10 +144,10 @@ class BaseAuthorizer:
         self.access_token: str | None = None
         self.scopes: set[str] | None = None
 
-    def _request_token(self, **data: Any):
+    async def _request_token(self, **data: Any):
         url = self._authenticator._requestor.reddit_url + const.ACCESS_TOKEN_PATH
         pre_request_time = time.time()
-        response = self._authenticator._post(url=url, **data)
+        response = await self._authenticator._post(url=url, **data)
         payload = response.json()
         if "error" in payload:  # Why are these OKAY responses?
             raise OAuthException(
@@ -184,37 +182,37 @@ class BaseAuthorizer:
             self.access_token is not None and time.time() < self._expiration_timestamp
         )
 
-    def revoke(self):
+    async def revoke(self):
         """Revoke the current Authorization."""
         if self.access_token is None:
             msg = "no token available to revoke"
             raise InvalidInvocation(msg)
 
-        self._authenticator.revoke_token(self.access_token, "access_token")
+        await self._authenticator.revoke_token(self.access_token, "access_token")
         self._clear_access_token()
 
 
-class TrustedAuthenticator(BaseAuthenticator):
+class AsyncTrustedAuthenticator(AsyncBaseAuthenticator):
     """Store OAuth2 authentication credentials for web, or script type apps."""
 
     RESPONSE_TYPE: str = "code"
 
     def __init__(
         self,
-        requestor: Requestor,
+        requestor: AsyncRequestor,
         client_id: str,
         client_secret: str,
         redirect_uri: str | None = None,
     ):
         """Represent a single authentication to Reddit's API.
 
-        :param requestor: An instance of :class:`.Requestor`.
+        :param requestor: An instance of :class:`.AsyncRequestor`.
         :param client_id: The OAuth2 client ID to use with the session.
         :param client_secret: The OAuth2 client secret to use with the session.
         :param redirect_uri: The redirect URI exactly as specified in your OAuth
             application settings on Reddit. This parameter is required if you want to
             use the :meth:`~.Authorizer.authorize_url` method, or the
-            :meth:`~.Authorizer.authorize` method of the :class:`.Authorizer` class
+            :meth:`~.Authorizer.authorize` method of the :class:`.AsyncAuthorizer` class
             (default: ``None``).
 
         """
@@ -225,37 +223,42 @@ class TrustedAuthenticator(BaseAuthenticator):
         return self.client_id, self.client_secret
 
 
-class UntrustedAuthenticator(BaseAuthenticator):
+class AsyncUntrustedAuthenticator(AsyncBaseAuthenticator):
     """Store OAuth2 authentication credentials for installed applications."""
 
     def _auth(self) -> tuple[str, str]:
         return self.client_id, ""
 
 
-class Authorizer(BaseAuthorizer):
+class AsyncAuthorizer(AsyncBaseAuthorizer):
     """Manages OAuth2 authorization tokens and scopes."""
 
     def __init__(
         self,
-        authenticator: BaseAuthenticator,
+        authenticator: AsyncBaseAuthenticator,
         *,
-        post_refresh_callback: Callable[[Authorizer], None] | None = None,
-        pre_refresh_callback: Callable[[Authorizer], None] | None = None,
+        post_refresh_callback: (
+            Callable[[AsyncAuthorizer], Awaitable[None]] | None
+        ) = None,
+        pre_refresh_callback: (
+            Callable[[AsyncAuthorizer], Awaitable[None]] | None
+        ) = None,
         refresh_token: str | None = None,
     ):
         """Represent a single authorization to Reddit's API.
 
-        :param authenticator: An instance of a subclass of :class:`.BaseAuthenticator`.
+        :param authenticator: An instance of a subclass of
+            :class:`.AsyncBaseAuthenticator`.
         :param post_refresh_callback: When a single-argument function is passed, the
             function will be called prior to refreshing the access and refresh tokens.
-            The argument to the callback is the :class:`.Authorizer` instance. This
+            The argument to the callback is the :class:`.AsyncAuthorizer` instance. This
             callback can be used to inspect and modify the attributes of the
-            :class:`.Authorizer`.
+            :class:`.AsyncAuthorizer`.
         :param pre_refresh_callback: When a single-argument function is passed, the
             function will be called after refreshing the access and refresh tokens. The
-            argument to the callback is the :class:`.Authorizer` instance. This callback
-            can be used to inspect and modify the attributes of the
-            :class:`.Authorizer`.
+            argument to the callback is the :class:`.AsyncAuthorizer` instance. This
+            callback can be used to inspect and modify the attributes of the
+            :class:`.AsyncAuthorizer`.
         :param refresh_token: Enables the ability to refresh the authorization.
 
         """
@@ -264,7 +267,7 @@ class Authorizer(BaseAuthorizer):
         self._pre_refresh_callback = pre_refresh_callback
         self.refresh_token = refresh_token
 
-    def authorize(self, code: str):
+    async def authorize(self, code: str):
         """Obtain and set authorization tokens based on ``code``.
 
         :param code: The code obtained by an out-of-band authorization request to
@@ -274,26 +277,26 @@ class Authorizer(BaseAuthorizer):
         if self._authenticator.redirect_uri is None:
             msg = "redirect URI not provided"
             raise InvalidInvocation(msg)
-        self._request_token(
+        await self._request_token(
             code=code,
             grant_type="authorization_code",
             redirect_uri=self._authenticator.redirect_uri,
         )
 
-    def refresh(self):
+    async def refresh(self):
         """Obtain a new access token from the refresh_token."""
         if self._pre_refresh_callback:
-            self._pre_refresh_callback(self)
+            await self._pre_refresh_callback(self)
         if self.refresh_token is None:
             msg = "refresh token not provided"
             raise InvalidInvocation(msg)
-        self._request_token(
+        await self._request_token(
             grant_type="refresh_token", refresh_token=self.refresh_token
         )
         if self._post_refresh_callback:
-            self._post_refresh_callback(self)
+            await self._post_refresh_callback(self)
 
-    def revoke(self, only_access: bool = False):
+    async def revoke(self, only_access: bool = False):
         """Revoke the current Authorization.
 
         :param only_access: When explicitly set to ``True``, do not evict the refresh
@@ -304,28 +307,28 @@ class Authorizer(BaseAuthorizer):
 
         """
         if only_access or self.refresh_token is None:
-            super().revoke()
+            await super().revoke()
         else:
-            self._authenticator.revoke_token(self.refresh_token, "refresh_token")
+            await self._authenticator.revoke_token(self.refresh_token, "refresh_token")
             self._clear_access_token()
             self.refresh_token = None
 
 
-class ImplicitAuthorizer(BaseAuthorizer):
+class AsyncImplicitAuthorizer(AsyncBaseAuthorizer):
     """Manages implicit installed-app type authorizations."""
 
-    AUTHENTICATOR_CLASS = UntrustedAuthenticator
+    AUTHENTICATOR_CLASS = AsyncUntrustedAuthenticator
 
     def __init__(
         self,
-        authenticator: UntrustedAuthenticator,
+        authenticator: AsyncUntrustedAuthenticator,
         access_token: str,
         expires_in: int,
         scope: str,
     ):
         """Represent a single implicit authorization to Reddit's API.
 
-        :param authenticator: An instance of :class:`.UntrustedAuthenticator`.
+        :param authenticator: An instance of :class:`.AsyncUntrustedAuthenticator`.
         :param access_token: The access_token obtained from Reddit via callback to the
             authenticator's ``redirect_uri``.
         :param expires_in: The number of seconds the ``access_token`` is valid for. The
@@ -343,7 +346,7 @@ class ImplicitAuthorizer(BaseAuthorizer):
         self.scopes = set(scope.split(" "))
 
 
-class ReadOnlyAuthorizer(Authorizer):
+class AsyncReadOnlyAuthorizer(AsyncAuthorizer):
     """Manages authorizations that are not associated with a Reddit account.
 
     While the ``"*"`` scope will be available, some endpoints simply will not work due
@@ -351,11 +354,11 @@ class ReadOnlyAuthorizer(Authorizer):
 
     """
 
-    AUTHENTICATOR_CLASS = TrustedAuthenticator
+    AUTHENTICATOR_CLASS = AsyncTrustedAuthenticator
 
     def __init__(
         self,
-        authenticator: BaseAuthenticator,
+        authenticator: AsyncBaseAuthenticator,
         scopes: list[str] | None = None,
     ):
         """Represent a ReadOnly authorization to Reddit's API.
@@ -367,15 +370,15 @@ class ReadOnlyAuthorizer(Authorizer):
         super().__init__(authenticator)
         self._scopes = scopes
 
-    def refresh(self):
+    async def refresh(self):
         """Obtain a new ReadOnly access token."""
         additional_kwargs = {}
         if self._scopes:
             additional_kwargs["scope"] = " ".join(self._scopes)
-        self._request_token(grant_type="client_credentials", **additional_kwargs)
+        await self._request_token(grant_type="client_credentials", **additional_kwargs)
 
 
-class ScriptAuthorizer(Authorizer):
+class AsyncScriptAuthorizer(AsyncAuthorizer):
     """Manages personal-use script type authorizations.
 
     Only users who are listed as developers for the application will be granted access
@@ -383,11 +386,11 @@ class ScriptAuthorizer(Authorizer):
 
     """
 
-    AUTHENTICATOR_CLASS = TrustedAuthenticator
+    AUTHENTICATOR_CLASS = AsyncTrustedAuthenticator
 
     def __init__(
         self,
-        authenticator: BaseAuthenticator,
+        authenticator: AsyncBaseAuthenticator,
         username: str | None,
         password: str | None,
         two_factor_callback: Callable | None = None,
@@ -395,7 +398,7 @@ class ScriptAuthorizer(Authorizer):
     ):
         """Represent a single personal-use authorization to Reddit's API.
 
-        :param authenticator: An instance of :class:`.TrustedAuthenticator`.
+        :param authenticator: An instance of :class:`.AsyncTrustedAuthenticator`.
         :param username: The Reddit username of one of the application's developers.
         :param password: The password associated with ``username``.
         :param two_factor_callback: A function that returns OTPs (One-Time Passcodes),
@@ -411,7 +414,7 @@ class ScriptAuthorizer(Authorizer):
         self._two_factor_callback = two_factor_callback
         self._username = username
 
-    def refresh(self):
+    async def refresh(self):
         """Obtain a new personal-use script type access token."""
         additional_kwargs = {}
         if self._scopes:
@@ -419,7 +422,7 @@ class ScriptAuthorizer(Authorizer):
         two_factor_code = self._two_factor_callback and self._two_factor_callback()
         if two_factor_code:
             additional_kwargs["otp"] = two_factor_code
-        self._request_token(
+        await self._request_token(
             grant_type="password",
             username=self._username,
             password=self._password,
@@ -427,7 +430,7 @@ class ScriptAuthorizer(Authorizer):
         )
 
 
-class DeviceIDAuthorizer(BaseAuthorizer):
+class AsyncDeviceIDAuthorizer(AsyncBaseAuthorizer):
     """Manages app-only OAuth2 for 'installed' applications.
 
     While the ``"*"`` scope will be available, some endpoints simply will not work due
@@ -435,18 +438,18 @@ class DeviceIDAuthorizer(BaseAuthorizer):
 
     """
 
-    AUTHENTICATOR_CLASS = (TrustedAuthenticator, UntrustedAuthenticator)
+    AUTHENTICATOR_CLASS = (AsyncTrustedAuthenticator, AsyncUntrustedAuthenticator)
 
     def __init__(
         self,
-        authenticator: BaseAuthenticator,
+        authenticator: AsyncBaseAuthenticator,
         device_id: str | None = None,
         scopes: list[str] | None = None,
     ):
         """Represent an app-only OAuth2 authorization for 'installed' apps.
 
-        :param authenticator: An instance of :class:`.UntrustedAuthenticator` or
-            :class:`.TrustedAuthenticator`.
+        :param authenticator: An instance of :class:`.AsyncUntrustedAuthenticator` or
+            :class:`.AsyncTrustedAuthenticator`.
         :param device_id: A unique ID (20-30 character ASCII string) (default:
             ``None``). ``device_id`` is set to ``"DO_NOT_TRACK_THIS_DEVICE"`` when the
             default argument is used. For more information about this parameter, see:
@@ -461,13 +464,13 @@ class DeviceIDAuthorizer(BaseAuthorizer):
         self._device_id = device_id
         self._scopes = scopes
 
-    def refresh(self):
+    async def refresh(self):
         """Obtain a new access token."""
         additional_kwargs = {}
         if self._scopes:
             additional_kwargs["scope"] = " ".join(self._scopes)
         grant_type = "https://oauth.reddit.com/grants/installed_client"
-        self._request_token(
+        await self._request_token(
             grant_type=grant_type,
             device_id=self._device_id,
             **additional_kwargs,